Penyesuaian Class Notifikasi
